+++ ios/Runner/GoogleService-Info.plist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyDztNijyxae_YAF47HTKy6kQIeDcdJlp2s</string>
+	<key>GCM_SENDER_ID</key>
+	<string>611673187318</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.example.puppycode</string>
+	<key>PROJECT_ID</key>
+	<string>pawpaw-1a214</string>
+	<key>STORAGE_BUCKET</key>
+	<string>pawpaw-1a214.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:611673187318:ios:c90073265483d45c349dd7</string>
+</dict>
+</plist>
